vector-db-tests/build-index/requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
faiss-cpu==1.7.4
filelock==3.12.3
fsspec==2023.9.0
huggingface-hub==0.16.4
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
packaging==23.1
pandas==2.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
six==1.16.0
sympy==1.12
tokenizers==0.13.3
torch==2.0.1
tqdm==4.66.1
transformers==4.33.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
